Fix all mismatched_lifetime_syntaxes warnings
Since Rust 1.89.0, a warning is emitted by default when the lifetime of
the return type is implicitly elided but matches an input parameter of
the function.

rustc tells us how to fix it (but weirdly enough there is no cargo fix
for it), so let’s do that, it makes the elided lifetime much more
evident anyway.